move RAMBlockMetadataStore to test package
